refactor _extract_strides, _unpack, etc; do not use torch.autograd.Function since we do not need backward()
